Transform invalid keys with a reasonable scheme.
There is a mismatch between what Hadoop supports for a key and
what Couchbase Server supports.  This change will convert as
follows:

\r  = ""
\n  = ""
\t  = "-"
" " = "_"

This change will emit warnings when doing this transformation.
The recommendation is to perform whatever kind of transformations
are needed inside Hadoop Map-Reduce jobs before export into Couchbase.

Change-Id: Ie20683746ce90b4e6c8cbe7246fcfc2909d0bea3
Reviewed-on: http://review.couchbase.org/18074
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>