Read proper fields from SqoopRecord to export.
The work for splitting the incoming record is done for us based
on the delimiter in the file, we simply need to extract the key
and value out of the incoming Key.

With this change, all built in tests pass and a full import
followed by an export can be run successfully.  The test run
included various datatypes and objects up to 10KByte in size.

Change-Id: I63caa8b28fbc3855d55061aa99e4c66c63f6319f
Reviewed-on: http://review.couchbase.org/18079
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>