Update dependencies to be compatible with CDH4.
I found that updating the dependencies to be compatible with CDH4
made the connector work correctly with no code changes. Note that
it won't work with CDH3 with these dependencies.  Also note that
compiling with CDH3 dependencies yeilds software that won't load
the proper classes under CDH4 owing to changes in CDH, not
anything we've done in the connector.

Note: the copyright was removed in one file here because the new
sqoop command does not allow # as a comment for some reason.

Change-Id: Ib19c1959d1817a202c53362f2cc2b8bd3ac9f9a1